app/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
cloudinary==1.41.0
distlib==0.3.9
Django==5.1.4
django-cors-headers==4.6.0
django-stubs==5.1.1
django-stubs-ext==5.1.1
django-types==0.19.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djangorestframework-stubs==3.15.1
djangorestframework-types==0.9.0
drf-yasg==1.21.8
drf-yasg-stubs==0.1.4
filelock==3.16.1
identify==2.6.3
idna==3.10
inflection==0.5.1
nodeenv==1.9.1
packaging==24.2
platformdirs==4.3.6
pre_commit==4.0.1
psycopg2==2.9.10
PyJWT==2.10.1
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.2
types-psycopg2==2.9.21.20241019
types-PyYAML==6.0.12.20240917
types-requests==2.32.0.20241016
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.28.0
